+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 * Customizada viewlet plone.analytics para ter uma div em torno de seu conteúdo,
   possibilitando mapeamento no Diazo (closes `#30`_).
   [rodfersou]
+* Movidas customizacoes nitf do portal brasil para portal modelo (closes `#26`_).
+  [rodfersou]
 * Removidas customizações da pasta overrides to tema, e movidas para brasil.gov.portal
   (closes `#19`_).
   [rodfersou]
@@ -45,5 +47,6 @@ Changelog
 .. _`#17`: https://github.com/plonegovbr/brasil.gov.portal/issues/17
 .. _`#19`: https://github.com/plonegovbr/brasil.gov.portal/issues/19
 .. _`#21`: https://github.com/plonegovbr/brasil.gov.portal/issues/21
+.. _`#26`: https://github.com/plonegovbr/brasil.gov.portal/issues/26
 .. _`#30`: https://github.com/plonegovbr/brasil.gov.portal/issues/30